support isolated process

+++ patch-loader/src/main/java/org/lsposed/lspatch/loader/LSPApplication.java
@@ -4,6 +4,7 @@ import static org.lsposed.lspatch.share.Constants.CONFIG_ASSET_PATH;
 import static org.lsposed.lspatch.share.Constants.ORIGINAL_APK_ASSET_PATH;
 
 import android.app.ActivityThread;
+import android.app.AppComponentFactory;
 import android.app.LoadedApk;
 import android.content.Context;
 import android.content.pm.ApplicationInfo;
@@ -14,6 +15,7 @@ import android.os.Build;
 import android.os.Parcel;
 import android.os.Parcelable;
 import android.system.Os;
+import android.util.ArrayMap;
 import android.util.Log;
 
 import com.google.gson.Gson;
@@ -21,6 +23,7 @@ import com.google.gson.Gson;
 import org.lsposed.lspatch.loader.util.FileUtils;
 import org.lsposed.lspatch.loader.util.XLog;
 import org.lsposed.lspatch.service.LocalApplicationService;
+import org.lsposed.lspatch.service.NullApplicationService;
 import org.lsposed.lspatch.service.RemoteApplicationService;
 import org.lsposed.lspatch.share.Constants;
 import org.lsposed.lspatch.share.PatchConfig;
@@ -44,10 +47,12 @@ import java.util.ArrayList;
 import java.util.Collections;
 import java.util.Map;
 import java.util.function.BiConsumer;
+import java.util.zip.ZipEntry;
 import java.util.zip.ZipFile;
 
-import de.robv.android.xposed.XC_MethodHook;
 import de.robv.android.xposed.XposedHelpers;
+import org.lsposed.lspatch.loader.util.FastZipEntry;
+import org.lsposed.lspatch.loader.util.FastZipIn;
 import hidden.HiddenApiBridge;
 
 /**
@@ -65,51 +70,89 @@ public class LSPApplication {
     private static LoadedApk appLoadedApk;
 
     private static PatchConfig config;
+    private static String cacheApk;
 
     public static boolean isIsolated() {
         return (android.os.Process.myUid() % PER_USER_RANGE) >= FIRST_APP_ZYGOTE_ISOLATED_UID;
     }
 
     public static void onLoad() {
-        if (isIsolated()) {
-            XLog.d(TAG, "Skip isolated process");
-            return;
-        }
         activityThread = ActivityThread.currentActivityThread();
-        var context = createLoadedApkWithContext();
-        if (context == null) {
-            XLog.e(TAG, "Error when creating context");
+
+        // we need a context for RemoteApplicationService or LocalApplicationService
+        var stubContext = createStubContext();
+        if (stubContext == null) {
+            XLog.e(TAG, "Error when creating stub context");
             return;
         }
-
+        loadConfig(stubContext);
         try {
             Log.d(TAG, "Initialize service client");
             ILSPApplicationService service;
-            if (config.useManager) {
-                service = new RemoteApplicationService(context);
+            if (isIsolated()) {
+                // not enable RemoteApplicationService in isolated process
+                // Caused by: java.lang.SecurityException: Isolated process not allowed to call getContentProvider
+                service = new NullApplicationService();
             } else {
-                service = new LocalApplicationService(context);
+                if (config.useManager) {
+                    service = new RemoteApplicationService(stubContext);
+                } else {
+                    service = new LocalApplicationService(stubContext);
+                }
+                disableProfile(stubContext);
             }
-
-            disableProfile(context);
             Startup.initXposed(false, ActivityThread.currentProcessName(), service);
             Log.i(TAG, "Bootstrap Xposed");
             Startup.bootstrapXposed();
-            // WARN: Since it uses `XResource`, the following class should not be initialized
-            // before forkPostCommon is invoke. Otherwise, you will get failure of XResources
-            Log.i(TAG, "Load modules");
-            LSPLoader.initModules(appLoadedApk);
-            Log.i(TAG, "Modules initialized");
+            Log.i(TAG, "Xposed initialized");
+
+            LoadedApk appLoadedApk = switchLoadedApk();
+            if (appLoadedApk == stubLoadedApk) {
+                Log.e(TAG, "appLoadedApk should diff with stubLoadedApk");
+            }
+            Log.i(TAG, "LoadedApk switched");
+
+            doSigBypass(stubContext);
+
+            // simulate instance appComponentFactory.
+            // this will give shells a chance to replace appComponentFactory again.
+            // and LoadedApkGetCLHooker will trigger Xposed modules.
+            appLoadedApk.getClassLoader();
 
             switchAllClassLoader();
-            doSigBypass(context);
         } catch (Throwable e) {
             throw new RuntimeException("Do hook", e);
         }
         Log.i(TAG, "LSPatch bootstrap completed");
     }
 
-    private static Context createLoadedApkWithContext() {
+    private static void loadConfig(Context stub) {
+        try (var is = stub.getClassLoader().getResourceAsStream(CONFIG_ASSET_PATH)) {
+            BufferedReader streamReader = new BufferedReader(new InputStreamReader(is, StandardCharsets.UTF_8));
+            config = new Gson().fromJson(streamReader, PatchConfig.class);
+        } catch (IOException e) {
+            Log.e(TAG, "Failed to load config file");
+        }
+        Log.i(TAG, "Use manager: " + config.useManager);
+        Log.i(TAG, "Signature bypass level: " + config.sigBypassLevel);
+    }
+
+    private static Context createStubContext() {
+        try {
+            var mBoundApplication = XposedHelpers.getObjectField(activityThread, "mBoundApplication");
+            var stubLoadedApk = (LoadedApk) XposedHelpers.getObjectField(mBoundApplication, "info");
+            var appInfo = (ApplicationInfo) XposedHelpers.getObjectField(mBoundApplication, "appInfo");
+
+            // reset appComponentFactory to prevent loop
+            appInfo.appComponentFactory = AppComponentFactory.class.getName();
+            return (Context) XposedHelpers.callStaticMethod(Class.forName("android.app.ContextImpl"), "createAppContext", activityThread, stubLoadedApk);
+        } catch (Throwable e) {
+            Log.e(TAG, "createStubContext", e);
+            return null;
+        }
+    }
+
+    private static LoadedApk switchLoadedApk() {
         try {
             var mBoundApplication = XposedHelpers.getObjectField(activityThread, "mBoundApplication");
 
@@ -117,36 +160,39 @@ public class LSPApplication {
             var appInfo = (ApplicationInfo) XposedHelpers.getObjectField(mBoundApplication, "appInfo");
             var compatInfo = (CompatibilityInfo) XposedHelpers.getObjectField(mBoundApplication, "compatInfo");
             var baseClassLoader = stubLoadedApk.getClassLoader();
+            Log.d(TAG, "stubLoadedApk.getClassLoader() " + baseClassLoader.hashCode());
 
-            try (var is = baseClassLoader.getResourceAsStream(CONFIG_ASSET_PATH)) {
-                BufferedReader streamReader = new BufferedReader(new InputStreamReader(is, StandardCharsets.UTF_8));
-                config = new Gson().fromJson(streamReader, PatchConfig.class);
-            } catch (IOException e) {
-                Log.e(TAG, "Failed to load config file");
-                return null;
-            }
-            Log.i(TAG, "Use manager: " + config.useManager);
-            Log.i(TAG, "Signature bypass level: " + config.sigBypassLevel);
-
+            String sourceDir = appInfo.sourceDir;
             Path originPath = Paths.get(appInfo.dataDir, "cache/lspatch/origin/");
-            Path cacheApkPath;
-            try (ZipFile sourceFile = new ZipFile(appInfo.sourceDir)) {
-                cacheApkPath = originPath.resolve(sourceFile.getEntry(ORIGINAL_APK_ASSET_PATH).getCrc() + ".apk");
-            }
-
-            appInfo.sourceDir = cacheApkPath.toString();
-            appInfo.publicSourceDir = cacheApkPath.toString();
-            appInfo.appComponentFactory = config.appComponentFactory;
-
-            if (!Files.exists(cacheApkPath)) {
-                Log.i(TAG, "Extract original apk");
-                FileUtils.deleteFolderIfExists(originPath);
-                Files.createDirectories(originPath);
-                try (InputStream is = baseClassLoader.getResourceAsStream(ORIGINAL_APK_ASSET_PATH)) {
-                    Files.copy(is, cacheApkPath);
+            try (ZipFile sourceFile = new ZipFile(sourceDir)) {
+                ZipEntry originalApkAsset = sourceFile.getEntry(ORIGINAL_APK_ASSET_PATH);
+                Path cacheApkPath = originPath.resolve(originalApkAsset.getCrc() + ".apk");
+                cacheApk = cacheApkPath.toString();
+                if(!isIsolated()) {
+                    if (!Files.exists(cacheApkPath)) {
+                        Log.i(TAG, "Extract original apk");
+                        FileUtils.deleteFolderIfExists(originPath);
+                        Files.createDirectories(originPath);
+                        try (InputStream is = sourceFile.getInputStream(originalApkAsset)) {
+                            Files.copy(is, cacheApkPath);
+                        }
+                    }
                 }
             }
+            appInfo.sourceDir = cacheApk;
+            appInfo.publicSourceDir = cacheApk;
+            replaceApplicationInfoCreator(appInfo.packageName);
+            clearCreatorCache();
 
+            if (config.appComponentFactory != null && config.appComponentFactory.length() > 0) {
+                if (config.appComponentFactory.startsWith(".")) {
+                    appInfo.appComponentFactory = appInfo.packageName + config.appComponentFactory;
+                } else {
+                    appInfo.appComponentFactory = config.appComponentFactory;
+                }
+            } else {
+                appInfo.appComponentFactory = null;
+            }
             var mPackages = (Map<?, ?>) XposedHelpers.getObjectField(activityThread, "mPackages");
             mPackages.remove(appInfo.packageName);
             appLoadedApk = activityThread.getPackageInfoNoCheck(appInfo, compatInfo);
@@ -171,17 +217,7 @@ public class LSPApplication {
                 }
             } catch (Throwable ignored) {}
             Log.i(TAG, "hooked app initialized: " + appLoadedApk);
-
-            var context = (Context) XposedHelpers.callStaticMethod(Class.forName("android.app.ContextImpl"), "createAppContext", activityThread, stubLoadedApk);
-            if (config.appComponentFactory != null) {
-                try {
-                    context.getClassLoader().loadClass(config.appComponentFactory);
-                } catch (ClassNotFoundException e) { // This will happen on some strange shells like 360
-                    Log.w(TAG, "Original AppComponentFactory not found: " + config.appComponentFactory);
-                    appInfo.appComponentFactory = null;
-                }
-            }
-            return context;
+            return appLoadedApk;
         } catch (Throwable e) {
             Log.e(TAG, "createLoadedApk", e);
             return null;
@@ -244,6 +280,10 @@ public class LSPApplication {
 
     private static void bypassSignature(Context context) {
         String packageName = context.getPackageName();
+        replacePackageInfoCreator(packageName);
+    }
+
+    private static void replacePackageInfoCreator(String packageName) {
         Parcelable.Creator<PackageInfo> originalCreator = PackageInfo.CREATOR;
         Parcelable.Creator<PackageInfo> proxiedCreator = new Parcelable.Creator<PackageInfo>() {
             @Override
@@ -273,6 +313,10 @@ public class LSPApplication {
             }
         };
         XposedHelpers.setStaticObjectField(PackageInfo.class, "CREATOR", proxiedCreator);
+        clearCreatorCache();
+    }
+
+    private static void clearCreatorCache() {
         try {
             Map<?, ?> mCreators = (Map<?, ?>) XposedHelpers.getStaticObjectField(Parcel.class, "mCreators");
             mCreators.clear();
@@ -289,17 +333,54 @@ public class LSPApplication {
         }
     }
 
+    private static void replaceApplicationInfoCreator(String packageName) {
+        Parcelable.Creator<ApplicationInfo> originalCreator = ApplicationInfo.CREATOR;
+        Parcelable.Creator<ApplicationInfo> proxiedCreator = new Parcelable.Creator<ApplicationInfo>() {
+            @Override
+            public ApplicationInfo createFromParcel(Parcel source) {
+                ApplicationInfo applicationInfo = originalCreator.createFromParcel(source);
+                if (applicationInfo != null && applicationInfo.packageName.equals(packageName)) {
+                    if (cacheApk != null) {
+                        if (config.sigBypassLevel >= Constants.SIGBYPASS_LV_PM) {
+                            applicationInfo.sourceDir = cacheApk;
+                            applicationInfo.publicSourceDir = cacheApk;
+                        }
+                    }
+                }
+                return applicationInfo;
+            }
+
+            @Override
+            public ApplicationInfo[] newArray(int size) {
+                return originalCreator.newArray(size);
+            }
+        };
+        XposedHelpers.setStaticObjectField(ApplicationInfo.class, "CREATOR", proxiedCreator);
+        clearCreatorCache();
+    }
+
     private static void doSigBypass(Context context) throws IllegalAccessException, ClassNotFoundException, IOException, NoSuchFieldException {
         if (config.sigBypassLevel >= Constants.SIGBYPASS_LV_PM) {
             XLog.d(TAG, "Original signature: " + config.originalSignature.substring(0, 16) + "...");
             bypassSignature(context);
         }
-        if (config.sigBypassLevel >= Constants.SIGBYPASS_LV_PM_OPENAT) {
-            String cacheApkPath;
-            try (ZipFile sourceFile = new ZipFile(context.getPackageResourcePath())) {
-                cacheApkPath = context.getCacheDir() + "/lspatch/origin/" + sourceFile.getEntry(ORIGINAL_APK_ASSET_PATH).getCrc() + ".apk";
+        if (config.sigBypassLevel >= Constants.SIGBYPASS_LV_PM_OPENAT && cacheApk != null) {
+            String patchedApk = context.getPackageResourcePath();
+
+            if (isIsolated()) {
+                try (FastZipIn in = new FastZipIn(new File(patchedApk))) {
+                    for (FastZipEntry entry : in.entries()) {
+                        if (ORIGINAL_APK_ASSET_PATH.equals(entry.name)) {
+                            long offset = in.getEntryDataStart(entry);
+                            long size = entry.unCompressSize;
+
+                            XLog.d(TAG, String.format("enableOpenatHook (%s, %s, %d, %d)", patchedApk, cacheApk, offset, size));
+                            SigBypass.enableOpenatHook(cacheApk, patchedApk, offset, size);
+                            break;
+                        }
+                    }
+                }
             }
-            SigBypass.enableOpenatHook(context.getPackageResourcePath(), cacheApkPath);
         }
     }
